Move prepare/parameter handling tests into `params.rs`
