+++ datafusion/sql/tests/cases/mod.rs
@@ -17,4 +17,5 @@
 
 mod collection;
 mod diagnostic;
+mod params;
 mod plan_to_sql;
